+++ controllers/action-routes.js
@@ -1,0 +1,32 @@
+const router = require('express').Router();
+const { User, Provider, Ticket } = require('../models');
+const auth = require('../utils/auth');
+
+router.get('/profile/:id', auth, async (req, res) => {
+  try {
+    res.render('profile');
+  } catch (err) {
+    console.log(err);
+    res.status(500).json(err);
+  }
+});
+
+router.get('/ticket/', auth, async (req, res) => {
+  try {
+    res.render('ticket');
+  } catch (err) {
+    console.log(err);
+    res.status(500).json(err);
+  }
+});
+
+router.get('/ticket/:id', auth, async (req, res) => {
+  try {
+    res.render('ticket');
+  } catch (err) {
+    console.log(err);
+    res.status(500).json(err);
+  }
+});
+
+module.exports = router;
